requirements.txt
langchain
openai
tiktoken
pypdf
qdrant-client
fastapi
python-dotenv
uvicorn
-e .